Fixed rebase problems.
